move css grid classnames to Grid
