+++ common/views/components/Grid/Grid.js
@@ -1,11 +1,21 @@
 // @flow
-import {cssGrid} from '../../../utils/classnames';
 import type {SizeMap} from '../../../utils/classnames';
 
 type Props = {|
   children: React.Node[] | React.Node,
   sizes: SizeMap
 |}
+
+export function cssGrid(sizes: SizeMap): string {
+  const base = 'css-grid__cell';
+  const modifierClasses = Object.keys(sizes).map(key => {
+    const size = sizes[key];
+    const modifier = key.replace(/([a-z])([A-Z])/g, '$1-$2').toLowerCase();
+    return `${base}--${modifier}${size}`;
+  });
+
+  return [base].concat(modifierClasses).join(' ');
+}
 
 const Grid = ({children, sizes}: Props) => (
   <div className='css-grid__container'>
